Move RESTful APIs related to templates to v1alpha3
