+++ pkg/kapis/devops/v1alpha3/register.go
@@ -20,7 +20,9 @@ package v1alpha3
 
 import (
 	"kubesphere.io/devops/pkg/client/git"
+	"kubesphere.io/devops/pkg/kapis/devops/v1alpha3/common"
 	"kubesphere.io/devops/pkg/kapis/devops/v1alpha3/scm"
+	"kubesphere.io/devops/pkg/kapis/devops/v1alpha3/template"
 	"net/http"
 
 	restful "github.com/emicklei/go-restful"
@@ -47,20 +49,22 @@ var GroupVersion = schema.GroupVersion{Group: api.GroupName, Version: "v1alpha3"
 // AddToContainer adds web service into container.
 func AddToContainer(container *restful.Container, devopsClient devopsClient.Interface,
 	k8sClient k8s.Client, client client.Client) (wss []*restful.WebService) {
-	ws := runtime.NewWebService(GroupVersion)
-	wss = append(wss, ws)
-	registerRoutes(devopsClient, k8sClient, client, ws)
-	pipelinerun.RegisterRoutes(ws, client)
-	pipeline.RegisterRoutes(ws, client)
-	container.Add(ws)
 
-	ws = runtime.NewWebServiceWithoutGroup(GroupVersion)
-	wss = append(wss, ws)
-	registerRoutes(devopsClient, k8sClient, client, ws)
-	pipelinerun.RegisterRoutes(ws, client)
-	pipeline.RegisterRoutes(ws, client)
-	container.Add(ws)
-	return
+	services := []*restful.WebService{
+		runtime.NewWebService(v1alpha3.GroupVersion),
+		runtime.NewWebServiceWithoutGroup(v1alpha3.GroupVersion),
+	}
+
+	for _, service := range services {
+		registerRoutes(devopsClient, k8sClient, client, service)
+		pipelinerun.RegisterRoutes(service, client)
+		pipeline.RegisterRoutes(service, client)
+		template.RegisterRoutes(service, &common.Options{
+			GenericClient: client,
+		})
+		container.Add(service)
+	}
+	return services
 }
 
 func registerRoutes(devopsClient devopsClient.Interface, k8sClient k8s.Client, client client.Client, ws *restful.WebService) {
