+++ pkg/kapis/devops/v1alpha3/template/template_handler.go
@@ -23,13 +23,13 @@ import (
 	"kubesphere.io/devops/pkg/api/devops/v1alpha3"
 	"kubesphere.io/devops/pkg/apiserver/query"
 	"kubesphere.io/devops/pkg/kapis"
-	kapisv1alpha1 "kubesphere.io/devops/pkg/kapis/devops/v1alpha1/common"
+	"kubesphere.io/devops/pkg/kapis/devops/v1alpha3/common"
 	resourcev1alpha3 "kubesphere.io/devops/pkg/models/resources/v1alpha3"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 )
 
 func (h *handler) handleQuery(request *restful.Request, response *restful.Response) {
-	devopsName := request.PathParameter(kapisv1alpha1.DevopsPathParameter.Data().Name)
+	devopsName := request.PathParameter(common.DevopsPathParameter.Data().Name)
 	commonQuery := query.ParseQueryParameter(request)
 
 	kapis.ResponseWriter{Response: response}.WriteEntityOrError(h.queryTemplate(devopsName, commonQuery))
@@ -37,7 +37,7 @@ func (h *handler) handleQuery(request *restful.Request, response *restful.Respon
 
 func (h *handler) queryTemplate(devopsName string, commonQuery *query.Query) (*api.ListResult, error) {
 	templateList := &v1alpha3.TemplateList{}
-	if err := h.genericClient.List(context.Background(),
+	if err := h.List(context.Background(),
 		templateList,
 		client.InNamespace(devopsName),
 		client.MatchingLabelsSelector{
@@ -49,7 +49,7 @@ func (h *handler) queryTemplate(devopsName string, commonQuery *query.Query) (*a
 }
 
 func (h *handler) handleGetTemplate(request *restful.Request, response *restful.Response) {
-	devopsName := request.PathParameter(kapisv1alpha1.DevopsPathParameter.Data().Name)
+	devopsName := request.PathParameter(common.DevopsPathParameter.Data().Name)
 	templateName := request.PathParameter(TemplatePathParameter.Data().Name)
 
 	kapis.ResponseWriter{Response: response}.WriteEntityOrError(h.getTemplate(devopsName, templateName))
@@ -57,14 +57,14 @@ func (h *handler) handleGetTemplate(request *restful.Request, response *restful.
 
 func (h *handler) getTemplate(devopsName, templateName string) (*v1alpha3.Template, error) {
 	template := &v1alpha3.Template{}
-	if err := h.genericClient.Get(context.Background(), client.ObjectKey{Namespace: devopsName, Name: templateName}, template); err != nil {
+	if err := h.Get(context.Background(), client.ObjectKey{Namespace: devopsName, Name: templateName}, template); err != nil {
 		return nil, err
 	}
 	return template, nil
 }
 
 func (h *handler) handleRenderTemplate(request *restful.Request, response *restful.Response) {
-	devopsName := request.PathParameter(kapisv1alpha1.DevopsPathParameter.Data().Name)
+	devopsName := request.PathParameter(common.DevopsPathParameter.Data().Name)
 	templateName := request.PathParameter(TemplatePathParameter.Data().Name)
 
 	kapis.ResponseWriter{Response: response}.WriteEntityOrError(h.renderTemplate(devopsName, templateName))
